File permssions and home directory changes
Change file permissions on the temp file before moving it to ensure
there is no time that the ssh keys file has the wrong access.

Remove the home directory when removing the user.

Rework the Dockerfile so that changes to the source files don't
require too long to rebuild the image when running locally.